Correction erreur nombre d'offre + ajout suppression offre

+++ protected/controllers/OffreEmploiController.php
@@ -15,7 +15,6 @@ class OffreEmploiController extends Controller
 	{
 		return array(
 			'accessControl', // perform access control for CRUD operations
-			'postOnly + delete', // we only allow deletion via POST request
 		);
 	}
 
@@ -29,6 +28,7 @@ class OffreEmploiController extends Controller
 	{
 		$user = Yii::app()->user;
 
+		// Si employe on donne acces a la liste d'offre, à une ofre en particulier et à la possibilité de candidater
 		if($user->getState('type') == 'employe')
 		{
 			return array(
@@ -41,9 +41,9 @@ class OffreEmploiController extends Controller
 			);
 		}
 
+		// Si employe on donne acces a la liste d'offre, à une ofre en particulier à la modification d'une offre et à la suppression
 		if($user->getState('type') == 'entreprise')
 		{
-
 			return array(
 					array('allow',
 						  'actions'=>['view', 'index', 'delete', 'update'],
@@ -133,15 +133,19 @@ class OffreEmploiController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		echo "test";
-	/*	$candidatures = Postuler::model()->FindByAttributes(array('id_offre_emploi'=>$id)); // On récupère les candidatures de l'offre
-		echo "test";
+		// On récupère les candidatures de l'offre
+		$candidatures = Postuler::model()->FindAll("id_offre_emploi LIKE '$id'");
 		
-		// Suppression des candidatures de l'offre
-		foreach($candidatures as $candidature)
+		
+		// Suppression des candidatures de l'offre si il y en a
+		if($candidatures != null)
 		{
-			$candidature->delete();
+			foreach($candidatures as $candidature)
+			{
+				$candidature->delete();
+			}
 		}
+		
 
 		// Récupération de l'offre
 		$offre = OffreEmploi::model()->FindByAttributes(array('id_offre_emploi'=>$id));
@@ -149,9 +153,8 @@ class OffreEmploiController extends Controller
 		// Suppression de l'offre
 		$offre->delete();
 
-		$this->loadModel($id)->delete();
-		$this->redirect('index');
-	*/
+		//$this->loadModel($id)->delete();
+		$this->redirect('index.php?r=offreEmploi/index');
 
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 		//if(!isset($_GET['ajax']))
